requirements.txt
aniso8601==3.0.2
argcomplete==1.9.3
astroid==2.2.5
Babel==2.6.0
boto3==1.9.3
botocore==1.12.3
braintree==3.54.0
certifi==2018.10.15
cfn-flip==1.0.3
chardet==3.0.4
Click==7.0
colorama==0.4.1
defusedxml==0.5.0
dj-database-url==0.5.0
Django==2.1
django-categories==1.6
django-category==2.0.1
django-cors-headers==2.4.0
django-debug-toolbar==1.11
django-environ==0.4.5
django-filter==2.0.0
django-graphql-jwt==0.1.10
django-graphql-social-auth==0.1.4
django-heroku==0.3.1
django-js-asset==1.2.1
django-mptt==0.9.1
django-object-tools==2.0.0
django-pandas==0.6.0
django-phonenumber-field==2.0.1
django-storages==1.7.1
docutils==0.14
durationpy==0.5
factory-boy==2.11.1
Faker==1.0.2
future==0.16.0
graphene==2.1.3
graphene-django==2.2.0
graphql-core==2.1
graphql-relay==0.4.5
gunicorn==19.9.0
hjson==3.0.1
idna==2.7
iso8601==0.1.12
isort==4.3.4
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.20.0
lazy-object-proxy==1.3.1
mccabe==0.6.1
numpy==1.16.2
oauthlib==2.1.0
pandas==0.24.1
phonenumbers==8.9.13
Pillow==6.0.0
placebo==0.8.2
promise==2.1
psycopg2==2.7.7
PyJWT==1.6.4
pylint==2.3.1
python-dateutil==2.6.1
python3-openid==3.1.0
pytz==2018.5
PyYAML==3.13
requests==2.20.0
requests-oauthlib==1.0.0
Rx==1.6.1
s3transfer==0.1.13
singledispatch==3.4.0.3
six==1.11.0
social-auth-app-django==3.1.0
social-auth-core==2.0.0
sqlparse==0.3.0
stripe==2.30.1
text-unidecode==1.2
typed-ast==1.3.5
typing==3.6.4
unicode-slugify==0.1.3
Unidecode==1.0.22
urllib3==1.23
whitenoise==4.1.2
wrapt==1.10.11
